requirements.txt
appdirs==1.4.3
argcomplete==1.8.2
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
beautifulsoup4==4.5.3
bs4==0.0.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
configparser==3.7.3
constantly==15.1.0
crayons==0.2.0
cryptography==2.3.1
cssselect==1.0.3
Cython==0.29.2
dj-database-url==0.5.0
Django==2.1
django-cors-headers==2.4.0
djangorestframework==3.9.1
docutils==0.14
docx2txt==0.6
EbookLib==0.15
google-speech==1.1.0
gTTS==2.0.3
gTTS-token==1.1.2
gunicorn==19.9.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
Kivy==1.10.1
Kivy-Garden==0.1.4
lxml==4.2.4
mysqlclient==1.4.2.post1
parsel==1.5.0
Pillow==5.2.0
pipenv==2018.11.26
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.3.1
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
PyPDF2==1.26.0
python-pptx==0.6.5
pyttsx==1.1
pyttsx3==2.7
pytz==2018.5
queuelib==1.5.0
requests==2.18.4
Scrapy==1.5.1
selenium==3.141.0
service-identity==17.0.0
six==1.10.0
SpeechRecognition==3.6.3
termcolor==1.1.0
tika==1.18
Twisted==18.7.0
urllib3==1.22
virtualenv==16.4.0
virtualenv-clone==0.5.1
w3lib==1.19.0
web-cache==1.1.0
webdriver-manager==1.7
xlrd==1.0.0
XlsxWriter==1.0.5
zope.interface==4.5.0
